feat: Add Confirmation Function and Constraint on Delete Button
- Implemented a confirmation function before deletion.
- Added a constraint to restrict delete functionality under specific conditions.

+++ src/Component/Applicant/MyHomeRequests.jsx
@@ -175,13 +175,16 @@ const MyEvents = () => {
               View
             </button>
           </Link>
-
-          <button
-            className="btn btn-danger btn-sm"
-            onClick={() => handleDelete(event.eventId)}
-          >
-            Delete
-          </button>
+          {event.confirmedAt == null ? (
+            <>
+              <button
+                className="btn btn-danger btn-sm"
+                onClick={() => handleDelete(event.eventId)}
+              >
+                Delete
+              </button>
+            </>
+          ) : null}
         </>
       ),
     })),
